requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==3.2.25
django-cors-headers==4.1.0
django-dotenv==1.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
